Adding dark vision, bullet reload, and resolution fixes
Dark Light Vision: Enemies now fade into the darkness when not seen. I believe this add a certain level of tension to hiding in the light for any period of time. I experimented with changing the hue of the enemy sprites, but it was both inconsistent and not quite the effect I was looking for. Added an extra lil' tutorial level for this.

Ammo Recharging: Over time instead of all at once. Forces the player to spend a second or so more in the light, stopping players from just dipping a toe in and out of the light. On the backend, it's actually a much larger number than the ammo count would imply (10 => 1) that I implemented to avoid decimals.

Further Resolution Tinkering: I originally made the game in an oddly sized room, which takes up the entire PlaySpace. The issue that this presents is that there's no 'extra' room to scale out the vertical to fit the rest of the room. To get rid of the black bars for this project, I would need to remake all my levels due to the entire level being the screen. Shoutout to this video for showing me the error of my ways. Resolution and Aspect Ratio Management for Game Maker - Part 2 (youtube.com).

QOL Changes:

Game starts in full screen now.
Swapped the text to a different font (Gamemaker's base font doesn't play well it seems.)
Recentered the death screen text.
Added an escape command to, well, escape.
A buncha debugging additions no one will ever see.